+++ src/main/java/hpbui/gamerportal/entity/Role.java
@@ -1,21 +1,14 @@
 package hpbui.gamerportal.entity;
 
+import javax.persistence.*;
 import java.util.Set;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.ManyToMany;
-import javax.persistence.Table;
 
 @Entity
 @Table(name = "gamer_portal.roles")
 public class Role {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+    private Long id;
 	
 	@Column(length=45, nullable = false)
 	private String roleName;
@@ -23,11 +16,11 @@ public class Role {
     @ManyToMany(mappedBy="roles")
     private Set<Account> accounts;
 
-	public int getId() {
+    public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+    public void setId(Long id) {
 		this.id = id;
 	}
