+++ src/main/java/hpbui/gamerportal/repository/AccountGameRepository.java
@@ -8,11 +8,11 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import java.util.List;
 
-public interface AccountGameRepository extends PagingAndSortingRepository<AccountGame, Integer> {
-    List<AccountGame> findAccountGamesByIdAccountAndIdGame(int idAccount, int idGame);
+public interface AccountGameRepository extends PagingAndSortingRepository<AccountGame, Long> {
+    List<AccountGame> findAccountGamesByIdAccountAndIdGame(Long idAccount, Long idGame);
 
     @Query(value = "SELECT * FROM gamer_portal.account_game WHERE idGame = ?1 GROUP BY idGame, idAccount",
             countQuery = "SELECT COUNT(*) FROM gamer_portal.account_game WHERE idGame = ?1 GROUP BY idGame, idAccount",
             nativeQuery = true)
-    Page<AccountGame> findDistinctIdAccountAndIdGameByIdGame(int idGame, Pageable pageable);
+    Page<AccountGame> findDistinctIdAccountAndIdGameByIdGame(Long idGame, Pageable pageable);
 }
