+++ src/main/java/hpbui/gamerportal/repository/RelationshipRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface RelationshipRepository extends PagingAndSortingRepository<Relationship, Integer> {
-    Relationship findByIdAccountFromAndIdAccountTo(int idAccountFrom, int idAccountTo);
+public interface RelationshipRepository extends PagingAndSortingRepository<Relationship, Long> {
+    Relationship findByIdAccountFromAndIdAccountTo(Long idAccountFrom, Long idAccountTo);
 
-    Page<Relationship> findByIdAccountFrom(int idAccountFrom, Pageable pageable);
+    Page<Relationship> findByIdAccountFrom(Long idAccountFrom, Pageable pageable);
 }
